move all activity log events to ActivityLogService
