+++ app/core/kafka/schemas/metadata.items.activity.avsc
@@ -1,0 +1,67 @@
+{
+   "mappings": {
+      "properties": {
+         "activity_type": {
+            "type": "keyword"
+         },
+         "activity_time": {
+            "type": "date",
+            "format": "yyyy-MM-dd HH:mm:ss.SSSSSS||epoch_second"
+         },
+         "item_id": {
+            "type": "keyword"
+         },
+         "item_type": {
+            "type": "keyword"
+         },
+         "item_name": {
+            "type": "text",
+            "fields": {
+               "keyword": {
+                  "type": "keyword"
+               }
+            }
+         },
+         "item_parent_path": {
+            "type": "text",
+            "analyzer": "path_analyzer",
+            "fields": {
+               "keyword": {
+                  "type": "keyword"
+               }
+            }
+         },
+         "container_code": {
+            "type": "keyword"
+         },
+         "container_type": {
+            "type": "keyword"
+         },
+         "user": {
+            "type": "text",
+            "fields": {
+               "keyword": {
+                  "type": "keyword"
+               }
+            }
+         },
+         "imported_from": {
+            "type": "keyword"
+         },
+         "changes": {
+            "type": "nested",
+            "properties": {
+               "item_property": {
+                  "type": "keyword"
+               },
+               "old_value": {
+                  "type": "keyword"
+               },
+               "new_value": {
+                  "type": "keyword"
+               }
+            }
+         }
+      }
+   }
+}
