+++ pyproject.toml
@@ -38,6 +38,7 @@ greenlet = "^1.1.2"
 fastapi-pagination = "^0.9.3"
 async-lru = "^1.0.3"
 aiokafka = "^0.7.2"
+fastavro = "^1.5.2"
 
 
 [tool.poetry.dev-dependencies]
